# Conflicts:
#	app/src/main/java/com/emproto/hoabl/di/HomeComponent.kt
#	app/src/main/java/com/emproto/hoabl/feature/home/adapters/InsightsAdapter.kt
#	app/src/main/java/com/emproto/hoabl/feature/home/views/fragments/HomeFragment.kt
#	app/src/main/java/com/emproto/hoabl/feature/portfolio/views/PortfolioSpecificProjectView.kt
#	app/src/main/java/com/emproto/hoabl/repository/HomeRepository.kt
#	networklayer/src/main/java/com/emproto/networklayer/feature/HomeDataSource.kt

+++ networklayer/src/main/java/com/emproto/networklayer/response/portfolio/ivdetails/InvestmentInformation.kt
@@ -11,7 +11,7 @@ data class InvestmentInformation(
     val inventoryBucket: String,
     val inventoryId: String,
     val isCompleted: Boolean,
-    val otherExpenses: Int,
+    val otherExpenses: Double,
     val ownedSince: String,
     val owners: String,
     val registryAmount: Int,
